+++ include/decompress_ext4.h
@@ -40,5 +40,5 @@ typedef struct _ext4_chunk_header {
 #define EXT4_CHUNK_TYPE_NONE		0xCAC3
 
 int write_compressed_ext4(char* img_base, unsigned int sector_base);
-int check_compress_ext4(char *img_base, unsigned long long parti_size);
+int check_compress_ext4(char *img_base, unsigned int parti_size);
